Refactor _check_invalid function in hk43.py to update default argument type

+++ hidrokit/contrib/taruma/hk43.py
@@ -75,7 +75,7 @@ def _have_invalid(array, check):
     return bool(_get_invalid(array, check=check))
 
 
-def _check_invalid(array, check=np.float):
+def _check_invalid(array, check=float):
     if _have_invalid(array, check=check):
         return _get_invalid(array, check=check)
     return None
